Create a http request module and utilize for query request to Genius
The request module currently functions well for GET and POST w/ urlencoded
form data in the body. It allows for custom headers. It works for my
purposes of requesting a query of the Genius API as well as the future
commit of requesting access token from Spotify via POST. It needs more
fleshing out and testing if it will be used in more general circumstances.