Revert "Refactor server URL assignment and logging in App.js"
This reverts commit cbb49d9cbfda0565289a37b9add027da519f8753.

The purpose of this revert is to clear the few, probably inconsequential,
changes that were made post production deployment. I need to get the dev
setup as close to the current prod set up as possible so that i can verify
that the changes are working before deploying. Currently, the ips/ports/cors
issues are complicating the situation and the extra changes are adding
unnecessary question marks.